pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.6.0",
    "numpy==2.2.4",
    "ta-lib>=0.6.3",
    "pybit>=2.4.0",
    "pandas>=1.3.0",
    "beaker>=1.11.0",
    "python-dotenv>=0.19.0",
]

[project.scripts]
mcp-server-demo = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
sources = ["src"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

